requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiohttp-session==2.12.0
aiosignal==1.3.1
aiostream==0.4.5
anthropic==0.3.2
anyio==3.7.0
appdirs==1.4.4
argilla==0.0.1
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
autopep8==2.0.2
backoff==2.2.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
click-log==0.4.0
clickhouse-connect==0.6.4
cohere==4.11.2
colorama==0.4.6
coloredlogs==15.0.1
construct==2.10.68
construct-typing==0.5.6
cookiecutter==1.7.3
create-aio-app==0.0.11
cryptography==41.0.1
ctransformers==0.2.11
dataclasses-json==0.5.8
decopatch==1.4.10
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.6
diskcache==5.6.1
distro==1.8.0
dnspython==2.3.0
docarray==0.21.1
docker==6.1.3
docstring-parser==0.15
docutils==0.20.1
dotty-dict==1.3.1
duckdb==0.8.1
ecdsa==0.18.0
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fake-useragent==1.1.3
fastapi==0.98.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.1
gotrue==1.0.2
grpcio==1.56.0
grpcio-health-checking==1.56.0
grpcio-reflection==1.56.0
grpcio-tools==1.56.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.15.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.1
invoke==1.7.3
jaraco.classes==3.2.3
jcloud==0.2.12
jina==3.15.2
jina-hubble-sdk==0.39.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.3.1
jsonalias==0.1.1
keyring==24.2.0
langchain==0.0.219
langchainplus-sdk==0.0.17
llama-cpp-python==0.1.68
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
makefun==1.15.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mini-lambda==2.2.3
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aiohttp-client==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.0
pinecone-client==2.2.2
pkginfo==1.9.6
portalocker==2.7.0
postgrest==0.10.6
posthog==3.0.1
poyo==0.5.0
prometheus-client==0.17.0
promisio==0.1.1
protobuf==4.23.3
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
pulsar-client==3.2.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
PyContracts==1.8.12
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pymongo==4.4.0
pypandoc==1.11
pyparsing==3.1.0
pypdf==3.12.0
pysrt==1.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-gitlab==3.15.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-semantic-release==7.33.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
qdrant-client==1.3.1
readme-renderer==40.0
realtime==1.0.0
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
solana==0.30.2
solders==0.18.1
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
storage3==0.5.2
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
text-unidecode==1.3
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
twine==3.8.0
typer==0.9.0
types-cachetools==5.3.0.5
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.7.12
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
valid8==5.1.2
validators==0.20.0
watchfiles==0.19.0
weaviate-client==3.21.0
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
wikipedia==1.4.0
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
